% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{get_th}
\alias{get_th}
\alias{get_th.default}
\title{Extract the threshold (aka breakpoints) value(s)}
\usage{
get_th(object, ...)

\method{get_th}{default}(object, ...)
}
\arguments{
\item{object}{An object of class \emph{seg_lm}}

\item{...}{currently unused}
}
\value{
A vector
}
\description{
Extract the threshold (aka breakpoints) value(s)
}
\examples{
data_thresh <- sim_thresh()
mod_seg <- seglm(formula = y~x, data = data_thresh, th_var = "x", nthresh =1)
get_th(mod_seg)
}
