% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{regime}
\alias{regime}
\alias{regime.default}
\title{Indicate the regime/segment in which each observation is}
\usage{
regime(object, ...)

\method{regime}{default}(object, ...)
}
\arguments{
\item{object}{An object of class \emph{seg_lm}}

\item{...}{currently unused}
}
\value{
A vector
}
\description{
Indicate the regime/segment in which each observation is
}
\examples{
data_thresh <- sim_thresh()
mod_seg <- seglm_lm(formula = y~x, data = data_thresh, th_var = "x", nthresh =1)
head(regime(mod_seg))
}
