% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seglm-package.R
\docType{package}
\name{seglm-package}
\alias{seglm-package}
\alias{seglm}
\title{seglm: a package for segmented regression, also known as threshold regression, piecewise regression, or regression with structural break}
\description{
The main function is seglm

This package computes segmented, also known as threshold regressions.
}
\author{
\strong{Maintainer}: Matthieu Stigler \email{Matthieu.Stigler@gmail.com}

}
\keyword{internal}
