% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seglm_search.R
\name{seglm_search_dynprog}
\alias{seglm_search_dynprog}
\alias{seglm_search_grid}
\alias{print.seglm_search}
\alias{deviance.seglm_search}
\title{Low level function to search the thresholds}
\usage{
seglm_search_dynprog(X, y, th_var, nthresh = 1, trim = 0.15, ...)

seglm_search_grid(X, y, th_var, nthresh = 1, trim = 0.15,
  iter = TRUE, trace = FALSE, max.iter = 3, return_details = FALSE)

\method{print}{seglm_search}(x, ...)

\method{deviance}{seglm_search}(object, ...)
}
\arguments{
\item{X}{matrix of regressors}

\item{y}{matrix of the response}

\item{th_var}{the threshold variable.}

\item{nthresh}{The number of threshold/breakpoints to estimate.}

\item{trim}{The minimum percentage of observations to keep in each regime.}

\item{...}{Further argumetns passed to the underlying breakpoints function}

\item{iter, max.iter, trace, return_details}{arguments to set the number of iterations, as well return_details}

\item{x}{object of class \emph{seglm_search}}

\item{object}{object of class \emph{seglm_search}}
}
\value{
An object of class "seglm_search" and "list"
}
\description{
Low level function to search the thresholds
}
\examples{
data_thresh <- sim_thresh()
X_inp <-  as.matrix(data_thresh[, "x", drop = FALSE])
y_inp <-  as.matrix(data_thresh[, "y"])
seglm_search_dynprog(X=X_inp, y=y_inp, th_var = X_inp)
seglm_search_grid(X=X_inp, y=y_inp, th_var = X_inp)
}
