% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentedReg-package.R
\docType{package}
\name{segmentedReg-package}
\alias{segmentedReg}
\alias{segmentedReg-package}
\title{segmentedReg: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
\keyword{internal}
