% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segreg.fit.R
\name{segreg_fit}
\alias{segreg_fit}
\alias{print.seg_reg}
\title{Low level function to estimate a segmented/threshold regression by pre-specifying the threshold values}
\usage{
segreg_fit(X, y, th_var, nthresh = 1, th_val)

\method{print}{seg_reg}(x, ...)
}
\arguments{
\item{X}{matrix of regressors}

\item{y}{matrix of the response}

\item{th_var}{The threshold variable, i.e. variable to segment over.}

\item{nthresh}{The number of threshold/breakpoints to estimate.}

\item{th_val}{The pre-specified threshold/breakpoints values.}

\item{x}{object of class \emph{seg_reg}}

\item{...}{unused}
}
\value{
An object of class "seg_reg" and "lm"
}
\description{
Low level function to estimate a segmented/threshold regression by pre-specifying the threshold values
}
\examples{
data_thresh <- sim_thresh()
X_inp <-  as.matrix(data_thresh[, "x", drop = FALSE])
y_inp <-  as.matrix(data_thresh[, "y"])
segreg_fit(X=X_inp, y=y_inp, th_var = X_inp, th_val = 0)
}
