% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segReg_search.R
\name{segReg_search_dynprog}
\alias{segReg_search_dynprog}
\alias{print.segreg_search}
\title{Low level function to search the thresholds}
\usage{
segReg_search_dynprog(X, y, th_var, th, nthresh = 1, trim = 0.15)

\method{print}{segreg_search}(x, ...)
}
\arguments{
\item{X}{matrix of regressors}

\item{y}{matrix of the response}

\item{th_var}{the threshold variable.}

\item{nthresh}{The number of threshold/breakpoints to estimate.}

\item{trim}{The minimum percentage of observations to keep in each regime.}

\item{x}{object of class \emph{segreg_search}}
}
\value{
An object of class "segreg_search" and "list"
}
\description{
Low level function to search the thresholds
}
\section{Methods (by generic)}{
\itemize{
\item \code{print}: Print method for \code{segreg_search} object
}}

\examples{
data_thresh <- sim_thresh()
X_inp <-  as.matrix(data_thresh[, "x", drop = FALSE])
y_inp <-  as.matrix(data_thresh[, "y"])
segReg_search_dynprog(X=X_inp, y=y_inp, th_var = X_inp)
}
