% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seglm-package.R
\docType{package}
\name{seglm-package}
\alias{seglm-package}
\alias{seglm}
\title{seglm: a package for segmented regression, also known as threshold regression, piecewise regression, or regression with structural break}
\description{
The main function is \code{\link{seglm_lm}}, which will automatically estimate the threshold points,
and return a modified \code{lm} object, with supplementary \code{seglm_lm} objects.
Usual \code{lm} methods such as \code{coef},  \code{summary}, \code{deviance} can be used on the output.
Methods specific to segmented regressions such as \code{\link{regime}}, \code{\link{get_th}} have also been implemented.

This package computes segmented, also known as threshold regressions.
}
\details{
Install with:

devtools::install_github("MatthieuStigler/seglm")
}
\examples{
data_thresh <- sim_thresh()
out <- seglm_lm(formula = y~x, data = data_thresh, th_var_name = "x", nthresh =1)
data_thresh$pred <- predict(out)
data_thresh_ord <-  data_thresh[order(data_thresh$x),]
plot(y~x, data=data_thresh)
lines(y_true ~ x, data = data_thresh_ord, col=1)
lines(pred ~ x, data=data_thresh_ord, col=2)
legend("topleft", col = 1:2, legend = c("True curve", "estimated"), lty=1)
}
\author{
\strong{Maintainer}: Matthieu Stigler \email{Matthieu.Stigler@gmail.com}

}
\keyword{internal}
