% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seglm-package.R
\docType{package}
\name{seglm-package}
\alias{seglm}
\alias{seglm-package}
\title{seglm: Segmented regression}
\description{
This package computes segmented, also known as threshold regressions.
}
\author{
\strong{Maintainer}: Matthieu Stigler \email{Matthieu.Stigler@gmail.com}

}
\keyword{internal}
