% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seglm.R
\name{seglm}
\alias{seglm}
\title{High level function to estimate a segmented/threshold regression}
\usage{
seglm(formula, data, th_var_name, nthresh = 1, th = NULL, ...)
}
\arguments{
\item{formula}{the usual formula for the linear model to estimate}

\item{data}{the data-frame containing the variables in the model.}

\item{th_var_name}{the name of the threshold variable. Using \emph{.time} will run a standard changepoint model.}

\item{nthresh}{The number of threshold/breakpoints to estimate.}

\item{th}{Optional. The threshold value(s), should be of length \code{nthresh}.}

\item{...}{further arguments passed to the lower-level \code{\link{segreg_search}} function}
}
\value{
An object of class "seglm" and "lm"
}
\description{
High level function to estimate a segmented/threshold regression
}
\examples{
data_thresh <- sim_thresh()
seglm(formula = y~x, data = data_thresh, th_var_name = "x", nthresh =1)
}
