% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seglm.R
\name{seglm}
\alias{seglm}
\title{Low level function to estimate a segmented/threshold regression by pre-specifying the threshold values}
\usage{
seglm(formula, data, th_var_name, nthresh = 1, ...)
}
\arguments{
\item{formula}{the usual formula for the linear model to estimate}

\item{data}{the data-frame containing the variables in the model.}

\item{th_var_name}{the name of the threshold variable. Using \emph{.time} will run a standard changepoint model.}

\item{nthresh}{The number of threshold/breakpoints to estimate.}

\item{...}{further arguments passed to lower level \emph{TODO} functions}
}
\value{
An object of class "seglm" and "lm"
}
\description{
Low level function to estimate a segmented/threshold regression by pre-specifying the threshold values
}
\examples{
data_thresh <- sim_thresh()
seglm(formula = y~x, data = data_thresh, th_var_name = "x", nthresh =1)
}
