% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segmentedReg-package.R
\docType{package}
\name{segmentedReg-package}
\alias{segmentedReg}
\alias{segmentedReg-package}
\title{segmentedReg: Segmented regression}
\description{
This package computes segmented, also known as threshold regressions.
}
\author{
\strong{Maintainer}: Matthieu Stigler \email{Matthieu.Stigler@gmail.com}

}
\keyword{internal}
