% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segReg_search.R
\name{segReg_search_dynprog}
\alias{segReg_search_dynprog}
\title{Low level function to search the thresholds}
\usage{
segReg_search_dynprog(X, y, th_var, th, nthresh = 1, trim = 0.15)
}
\arguments{
\item{X}{matrix of regressors}

\item{y}{matrix of the response}

\item{th_var}{the threshold variable.}

\item{nthresh}{The number of threshold/breakpoints to estimate.}

\item{trim}{The minimum percentage of observations to keep in each regime.}
}
\value{
An object of class "segreg_search" and "list"
}
\description{
Low level function to search the thresholds
}
\examples{
add(1, 1)
}
