% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segReg_search.R
\name{segReg_search_dynprog}
\alias{segReg_search_dynprog}
\alias{segReg_search_grid}
\alias{print.segreg_search}
\alias{deviance.segreg_search}
\title{Low level function to search the thresholds}
\usage{
segReg_search_dynprog(X, y, th_var, nthresh = 1, trim = 0.15)

segReg_search_grid(X, y, th_var, nthresh = 1, trim = 0.15,
  iter = TRUE, trace = FALSE, max.iter = 3, return_details = FALSE)

\method{print}{segreg_search}(x, ...)

\method{deviance}{segreg_search}(object, ...)
}
\arguments{
\item{X}{matrix of regressors}

\item{y}{matrix of the response}

\item{th_var}{the threshold variable.}

\item{nthresh}{The number of threshold/breakpoints to estimate.}

\item{trim}{The minimum percentage of observations to keep in each regime.}

\item{iter, max.iter, trace, return_details}{arguments to set the number of iterations, as well return_details}

\item{x}{object of class \emph{segreg_search}}

\item{...}{unused}

\item{object}{object of class \emph{segreg_search}}
}
\value{
An object of class "segreg_search" and "list"
}
\description{
Low level function to search the thresholds
}
\examples{
data_thresh <- sim_thresh()
X_inp <-  as.matrix(data_thresh[, "x", drop = FALSE])
y_inp <-  as.matrix(data_thresh[, "y"])
segReg_search_dynprog(X=X_inp, y=y_inp, th_var = X_inp)
segReg_search_grid(X=X_inp, y=y_inp, th_var = X_inp)
}
