% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat.R
\name{sim_thresh}
\alias{sim_thresh}
\title{Simulation of data for example}
\usage{
sim_thresh(N = 200, th_val = 0, output = c("df", "matrix"))
}
\arguments{
\item{N}{Size of sample to simulate}

\item{th_val}{The pre-specified threshold/breakpoints values.}

\item{output}{whether to return a \emph{data.frame} or \emph{matrix}}
}
\value{
A data frame with columns \emph{y}, \emph{x1} and \emph{x2}
}
\description{
Simulation of data for example
}
\examples{
data_thresh <- sim_thresh()
}
